Refactor .htaccess and config files
This commit refactors the .htaccess file and the config.php file. In the .htaccess file, the code adds rewrite rules to redirect all requests to the index.php file. In the config.php file, the code updates the base_url based on the server's hostname. If the hostname is 'localhost', the base_url is set to 'http://localhost/perpustakan/Perpustakan/'. If the hostname is 'anggota-library', the base_url is set to 'http://anggota-library.test/'. Additionally, the code updates the sess_save_path in the config.php file to use the system's temporary directory.

The commit also updates the database.php file by changing the hostname from 'localhost' to '127.0.0.1'.

Finally, the index.php file is modified to set the environment variable based on the server's hostname. If the hostname is 'localhost' or 'anggota-library.test', the environment is set to 'development'. Otherwise, the environment is set to 'production'.